refactor: split main.rs into modules
One 2,200-line file becomes a module per concern: theme, menu, storage,
cursor, edit (pure text ops + tests), markdown, find, win95. Behaviour is
unchanged; the inline layouter's seven copy-pasted spans collapse into a
macro, dirty-marking gets a single helper, and the deprecated
allocate_ui_at_rect call and all clippy warnings are gone.

+++ src/mac_ime.rs
@@ -78,8 +78,7 @@ unsafe extern "C" fn insert_text(this: &NSObject, sel: Sel, string: &NSObject, r
 
     if range.location != NSNotFound as NSUInteger {
         // An accent picked from the panel, replacing the character it opened on.
-        STATE.lock().unwrap().pending =
-            Some((range.location as usize, range.length as usize, text));
+        STATE.lock().unwrap().pending = Some((range.location, range.length, text));
         PANEL_REPLACED.store(true, Ordering::Relaxed);
         if let Some(ctx) = REPAINT.lock().unwrap().as_ref() {
             ctx.request_repaint();
